requirements.txt
kafka-python==1.4.2
numpy==1.14.2
opencv-python==3.4.0.12
pika==0.11.2
